Refactored package structure
Improved the package structure so that the controllers and views have their
own packages and the models are moved outside the gui package.
