+++ src/main/java/gui/views/phylogeny/NewickColour.java
@@ -1,4 +1,4 @@
-package gui.phylogeny.model;
+package gui.views.phylogeny;
 
 import java.util.HashMap;
 import java.util.Map;
