+++ src/main/java/gui/views/phylogeny/NewickNode.java
@@ -1,16 +1,16 @@
-package gui.phylogeny.model;
+package gui.views.phylogeny;
 
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
+
+import gui.eventhandlers.phylogeny.NewickNodeMouseEventHandler;
 import javafx.scene.Group;
 import javafx.scene.control.Label;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.shape.Rectangle;
-
-import gui.phylogeny.eventhandlers.NewickNodeMouseEventHandler;
 
 /**
  * Class that is used to visualize a single node of a {@link NewickTree}
