+++ src/main/resources/splashScreen.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox fx:id="verticalBox" alignment="CENTER" minHeight="-Infinity" minWidth="-Infinity" prefHeight="535.0" prefWidth="800.0" style="-fx-background-color: lightgrey;" xmlns="http://javafx.com/javafx/8.0.71" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.SplashController">
+<VBox fx:id="verticalBox" alignment="CENTER" minHeight="-Infinity" minWidth="-Infinity" prefHeight="535.0" prefWidth="800.0" style="-fx-background-color: lightgrey;" xmlns="http://javafx.com/javafx/8.0.71" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.controllers.SplashController">
    <children>
       <ImageView fitHeight="500.0" fitWidth="800.0" pickOnBounds="true" preserveRatio="true">
          <image>
